refactor directory structure to use 'modules' namespace
