+++ geoparser/geoparserv2/modules/__init__.py
@@ -1,4 +1,3 @@
-from geoparser.geoparserv2.geoparserv2 import GeoparserV2
 from geoparser.geoparserv2.modules.interfaces import (
     BaseModule,
     RecognitionModule,
