+++ tests/geoparserv2/modules/test_interfaces.py
@@ -3,7 +3,7 @@ import json
 
 import pytest
 
-from geoparser.geoparserv2.module_interfaces import (
+from geoparser.geoparserv2.modules.interfaces import (
     BaseModule,
     RecognitionModule,
     ResolutionModule,
